Re: PR13802, referred common symbol not UNDefined
It looks like this one was previously just swept under the rug.  Prior
to git commit f667bf9efddd, the arm backend emitted an absolute
dynamic symbol "foo" with value zero for what was originally a common
symbol.  That was just wrong.

The patch also removes an xfail for bfin-linux-uclibc.

	PR 13802
	* testsuite/ld-elf/comm-data.exp: Don't xfail arm or bfin-linux.